requirements.txt
python-dotenv
requests
beautifulsoup4
lxml
schedule
openai
pandas
streamlit
